place minimax fun in utilz
