+++ React/src/app/utils/game-mode.ts
